% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_prepare_for_fitting.R
\name{MDfitDataSet}
\alias{MDfitDataSet}
\title{Creat a S3 class MDfitDataSet to store data matrix}
\usage{
MDfitDataSet(P, M, length.P = c(10000), length.M = c(1000), zt = zt,
  mode = "NB", fitType.dispersion = "local",
  fitType.var = "separate")
}
\arguments{
\item{P}{data for pre-mRNA}

\item{M}{data for mRNA}

\item{length.P}{intron length}

\item{length.M}{exon length}

\item{zt}{time points}

\item{mode}{"NB" or "logNormal"}

\item{fitType.var}{"separate" fit variances separate for pre-mRNA and mRNA}

\item{fitType.disperions}{parameter of dispersion fitting type in DESEq2}
}
\value{
A vector of nb of time points, mean of data, amplitude and relative amplitude of fitted result,  phase and pvalue
}
\description{
Creat a S3 class MDfitDataSet to store data matrix
}
\details{
Creat a S3 class MDfitDataSet to store data matrix, (P and M for pre-mRNA and mRNA), time points (zt), length.pre-mRNA and length.mRNA
scaling factors for rpkm calculation, dispersion parameters
some simple functions associated to extract these parameters, which will be used as global parameters in the model fitting
}
\author{
Jingkui Wang
}
